Move job watcher component impl to mixin
This didn’t belong in the job-detail-body anymore, so it’s been moved
up to the job-detail component and applied as a mixin

Actually it may work better to have it as a standalone component
