+++ app/components/job-detail.js
@@ -1,6 +1,7 @@
 import Ember from 'ember';
+import JobWatcher from 'bespin-ui/mixins/job-watcher';
 
-const JobDetail = Ember.Component.extend({
+const JobDetail = Ember.Component.extend(JobWatcher, {
   tagName: 'div',
   classNames: ['panel'],
   classNameBindings: 'panelType',
